Revert "Create flashcards list"
This reverts commit 3717862d04cb02907910ad0afcc40aaf98a682d6.
