+++ src/App.css
@@ -13,17 +13,6 @@ body {
 .cards_container {
   display: flex;
   flex-direction: column;
-  gap: 12px;
   align-items: center;
-  max-height: 695px;
-  padding-top: 12px;
-  padding-bottom: 12px;
-  overflow: auto;
-}
-.cards_container::-webkit-scrollbar {
-  width: 0.5em; /* Szerokość paska przewijania */
-  background-color: transparent; /* Tło paska przewijania */
-}
-.cards_container::-webkit-scrollbar-thumb {
-  background-color: #888; /* Kolor paska przewijania */
+  margin-top: 15px;
 }
