+++ src/components/Card.module.css
@@ -3,13 +3,12 @@
   flex-direction: column;
   background-color: white;
   width: 343px;
-  position: relative;
   min-height: 188px;
+  position: relative;
   border-radius: 15px;
   border: 0.5px #ebebec solid;
   justify-content: space-between;
   align-items: center;
-  flex-grow: 1;
 }
 .corner_wrapper {
   display: flex;
@@ -27,12 +26,7 @@
   cursor: pointer;
 }
 .text_wrapper {
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
+  position: relative;
 }
 .question_text {
   position: relative;
@@ -77,22 +71,4 @@
   word-wrap: break-word;
   font-weight: 500;
   cursor: pointer;
-}
-.output {
-  text-align: center;
-  justify-content: center;
-
-  border: none;
-  color: #161616;
-  font-size: 16px;
-  width: 315px;
-  font-weight: 500;
-  resize: none;
-  overflow: hidden;
-  line-height: 20px;
-  overflow-wrap: break-word;
-  padding: 12px 10px 10px 10px;
-  &:focus {
-    outline: none;
-  }
 }
